Updated index and home html page

+++ src/pipeline/predict_pipeline.py
@@ -28,8 +28,8 @@ class CustomData:
                  parental_level_of_education,
                  lunch:str,
                  test_preparation_course,
-                 reading_score:int,
-                 writing_score:int):
+                 reading_score:float,
+                 writing_score:float):
         
         self.gender = gender
         self.race_ethnicity=race_ethnicity
